feat: add prelude and update dependencies
