+++ client-grpc/examples/flightplan.rs
@@ -1,10 +1,7 @@
 //! Simulates a flow of ADS-B with multiple reporters
 use futures_lite::stream::StreamExt;
 use lib_common::grpc::get_endpoint_from_env;
-use svc_compliance_client_grpc::client::{FlightPlanRequest, RpcServiceClient};
-use svc_compliance_client_grpc::service::Client as ServiceClient;
-use svc_compliance_client_grpc::{Client, GrpcClient};
-use tonic::transport::Channel;
+use svc_compliance_client_grpc::prelude::{compliance::*, *};
 
 async fn mq_listener() -> Result<(), ()> {
     let mq_addr = format!("amqp://rabbitmq:5672");
@@ -57,8 +54,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     tokio::spawn(mq_listener());
 
     let (host, port) = get_endpoint_from_env("SERVER_HOSTNAME", "SERVER_PORT_GRPC");
-    let connection = GrpcClient::<RpcServiceClient<Channel>>::new_client(&host, port, "compliance");
-    let response = connection
+    let client = ComplianceClient::new_client(&host, port, "compliance");
+    let response = client
         .submit_flight_plan(FlightPlanRequest {
             flight_plan_id: "123".to_string(),
             data: r#"{ "test": 1 }"#.to_string(),
